pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.1",
    "discord-py>=2.4.0",
    "mistralai>=1.4.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.95.0",
    "uvicorn>=0.21.0",
    "cryptography>=40.0.0",
]
